refactor: Folder organization
